+++ backend/collaboration/package.json
@@ -2,17 +2,23 @@
   "name": "collaboration",
   "version": "1.0.0",
   "description": "",
-  "main": "server.js",
+  "main": "dist/server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon server.js"
+    "build": "rimraf dist && npx tsc",
+    "prestart": "npm run build",
+    "start": "npx tsc -w & nodemon dist/server.js"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/node": "^22.5.5",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.21.0",
-    "nodemon": "^3.1.4"
+    "nodemon": "^3.1.4",
+    "rimraf": "^6.0.1",
+    "typescript": "^5.6.2"
   }
 }
