+++ backend/matching/src/routes/matchingRoutes.ts
@@ -1,4 +1,4 @@
-const express = require('express');
+import express from 'express';
 
 const router = express.Router();
 
@@ -6,4 +6,4 @@ router.get('/', (req, res) => {
     res.send('Hello from matching service!');
 });
 
-module.exports = router;
+export default router;
