+++ backend/user/.dockerignore
@@ -1,1 +1,4 @@
 .env
+node_modules/
+dist/
+sample.env
